chore: remove unnecessary functionality
