+++ package.json
@@ -21,7 +21,6 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.7",
-    "@testing-library/user-event": "^14.5.1",
     "@types/react-redux": "^7.1.30",
     "axios-mock-adapter": "^1.22.0",
     "query-string": "^8.1.0",
@@ -32,6 +31,7 @@
     "react-router-dom": "^6.18.0"
   },
   "devDependencies": {
+    "@testing-library/user-event": "^14.5.1",
     "@testing-library/jest-dom": "^6.1.4",
     "@testing-library/react": "^14.1.0",
     "@types/history": "^5.0.0",
