+++ src/services/getPagesCount.ts
@@ -1,7 +1,3 @@
 export function getPagesArray(totalPages: number) {
-  const pagesArray: string[] = [];
-  for (let i = 1; i <= totalPages; i++) {
-    pagesArray.push(i.toString());
-  }
-  return pagesArray;
+  return Array.from({ length: totalPages }, (_e, i) => (i + 1).toString());
 }
